Clean up WM package
