+++ wm/switcher.go
@@ -14,6 +14,43 @@ import (
 // Therefore the UI cannot handle keyboard input - so we add it here instead.
 var switcherInstance *ui.Switcher
 
+func (x *x11WM) applyAppSwitcher() {
+	if switcherInstance == nil {
+		return
+	}
+
+	go switcherInstance.HideApply()
+	xproto.UngrabKeyboard(x.x.Conn(), xproto.TimeCurrentTime)
+	windowClientListStackingUpdate(x)
+	switcherInstance = nil
+}
+
+func (x *x11WM) cancelAppSwitcher() {
+	if switcherInstance == nil {
+		return
+	}
+
+	go switcherInstance.HideCancel()
+	xproto.UngrabKeyboard(x.x.Conn(), xproto.TimeCurrentTime)
+	switcherInstance = nil
+}
+
+func (x *x11WM) nextAppSwitcher() {
+	if switcherInstance == nil {
+		return
+	}
+
+	go switcherInstance.Next()
+}
+
+func (x *x11WM) previousAppSwitcher() {
+	if switcherInstance == nil {
+		return
+	}
+
+	go switcherInstance.Previous()
+}
+
 func (x *x11WM) showOrSelectAppSwitcher(reverse bool) {
 	xproto.GrabKeyboard(x.x.Conn(), true, x.x.RootWin(), xproto.TimeCurrentTime, xproto.GrabModeAsync, xproto.GrabModeAsync)
 
@@ -34,41 +71,4 @@ func (x *x11WM) showOrSelectAppSwitcher(reverse bool) {
 			switcherInstance = ui.ShowAppSwitcher(x.Windows(), desktop.Instance().IconProvider())
 		}
 	}()
-}
-
-func (x *x11WM) previousAppSwitcher() {
-	if switcherInstance == nil {
-		return
-	}
-
-	go switcherInstance.Previous()
-}
-
-func (x *x11WM) nextAppSwitcher() {
-	if switcherInstance == nil {
-		return
-	}
-
-	go switcherInstance.Next()
-}
-
-func (x *x11WM) cancelAppSwitcher() {
-	if switcherInstance == nil {
-		return
-	}
-
-	go switcherInstance.HideCancel()
-	xproto.UngrabKeyboard(x.x.Conn(), xproto.TimeCurrentTime)
-	switcherInstance = nil
-}
-
-func (x *x11WM) applyAppSwitcher() {
-	if switcherInstance == nil {
-		return
-	}
-
-	go switcherInstance.HideApply()
-	xproto.UngrabKeyboard(x.x.Conn(), xproto.TimeCurrentTime)
-	windowClientListStackingUpdate(x)
-	switcherInstance = nil
 }
